Add mime-type lookup to Response.type(...)
The documentation said it was supposed to do mime type lookup for you
(e.g. `type('.html') => 'text/html'`), but that was unimplemented.

Rather than add a dependency on the `mime` package, we have pulled in
the "lite" database used by that package. We did this to reduce the
package size. The `mime` package, even though it doesn't have any
dependencies, is fairly large because of the additional more
full-featured database that it includes.  Since it's those additional
(rarely-used) types are unlikely to be needed, and we're trying to keep
our `@silvermine/lambda-express` package size small, we just imported
the "lite" database and wrote our own wrapper for it. The MIT licensing
made this acceptable to do.