Fix mime type lookup not having standard.json available
When building the `dist/esm` folder, the file `src/mime/standard.json`
was not getting copied into `dist/esm/mime/`. Thus, when someone
imported the module in their node project, Lambda Express could not be
initialized because it was missing a file that was being required. The
reason the file wasn't being copied into the dist was that it was being
required instead of imported.

Importing it also has the side effect of providing type inspection of
standard.json and thus making the `reduce` loop type safe.