Revert path-to-regexp to same version Express uses
There were a few differences between the version Express uses (0.1.7)
and the version we started out using (3.0.0). Notably, the support for
`/*` was drastically different. Since that one pattern will be used so
frequently, it didn't seem worth trying to maintain that difference as
it would negatively impact developers familiar with the `/*` pattern.

There was also a difference in the support of modifiers indicating how
many times a named parameter could appear (e.g. `/:path+` for params
like `/some/directory/and/file/path`).

The differences were clearly explained on the [path-to-regexp
homepage][1], but I originally thought it might be worth accepting those
differences or working around them. This doesn't seem feasible or worth
it, though.

[1]: https://github.com/pillarjs/path-to-regexp/blob/4eee1e15ba72d93c996bac4ae649a846eb326562/Readme.md#compatibility-with-express--4x