Add tests for routes without slashes
Previously, we'd noted that routes that ended in a slash were treated
differently from routes that did not end in a slash. For example, if a
route handler was registered for the GET method for the pattern
`/cars/`, and then a GET request is made for `/cars`, then there would
be no match. Express, however, would treat it as a match. This was fixed
in 2799ecf where the version of `path-to-regexp` was downgraded to match
Express'.