fix: decode + as ' ' in query string parameters
While the events that come from APIGW already have the query string
parameters decoded, ALB events do not auto-decode query string
parameters. For most cases this is OK since we decode the parameters
upon Request construction. However, this was failing for the case of
a '+'. The '+' in query string parameters from ALB events were not
being treated as a space, as they should have been. As noted in the
MDN docs for `decodeURIComponent` (which is what we were using to
decode the query string), '+' is not handled as you might expect.
We now do the "preparation" mentioned in the docs.